Revert "User telemetry title added resource name and tab name"
This reverts commit 62500a052c88797fc025ff4d4bcc9cb8461ee7fc.
