+++ frontend/packages/ceph-storage-plugin/src/components/dashboards/block-pool/block-pool-dashboard.tsx
@@ -1,10 +1,7 @@
 import * as React from 'react';
-import Helmet from 'react-helmet';
-import { useTranslation } from 'react-i18next';
 
 import Dashboard from '@console/shared/src/components/dashboard/Dashboard';
 import DashboardGrid from '@console/shared/src/components/dashboard/DashboardGrid';
-import { labelForNodeKind, labelKeyForNodeKind } from '@console/shared';
 
 import { StoragePoolKind } from 'packages/ceph-storage-plugin/src/types';
 import { BlockPoolDashboardContext } from './block-pool-dashboard-context';
@@ -25,22 +22,12 @@ const mainCards = [{ Card: StatusCard }, { Card: RawCapacityCard }, { Card: Util
 const rightCards = [{ Card: MirroringCard }];
 
 export const BlockPoolDashboard: React.FC<PoolDashboardProps> = ({ obj }) => {
-  const { t } = useTranslation();
   return (
-    <>
-      <Helmet>
-        <title data-title-id={`${labelForNodeKind(obj.kind)} · Overview`}>
-          {obj.metadata.name}
-          {' · '} {t(labelKeyForNodeKind(obj.kind))}
-          {' · '} {t('ceph-storage-plugin~Overview')}
-        </title>
-      </Helmet>
-      <BlockPoolDashboardContext.Provider value={{ obj }}>
-        <Dashboard>
-          <DashboardGrid mainCards={mainCards} leftCards={leftCards} rightCards={rightCards} />
-        </Dashboard>
-      </BlockPoolDashboardContext.Provider>
-    </>
+    <BlockPoolDashboardContext.Provider value={{ obj }}>
+      <Dashboard>
+        <DashboardGrid mainCards={mainCards} leftCards={leftCards} rightCards={rightCards} />
+      </Dashboard>
+    </BlockPoolDashboardContext.Provider>
   );
 };
