+++ frontend/packages/console-app/locales/en/console-app.json
@@ -250,10 +250,10 @@
   "Name": "Name",
   "Version": "Version",
   "Status": "Status",
-  "Console plugins": "Console plugins",
   "Console operator spec.managementState is unmanaged. Changes to plugins will have no effect.": "Console operator spec.managementState is unmanaged. Changes to plugins will have no effect.",
   "Console plugins table": "Console plugins table",
   "console plugins": "console plugins",
+  "Console plugins": "Console plugins",
   "Customize": "Customize",
   "Updating cluster to {{version}}": "Updating cluster to {{version}}",
   "API Servers": "API Servers",
