+++ frontend/packages/console-app/src/components/pdb/PDBDetailsPage.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import * as _ from 'lodash';
-import Helmet from 'react-helmet';
 import { useTranslation } from 'react-i18next';
 import { DetailsPage } from '@console/internal/components/factory';
 import {
@@ -10,7 +9,6 @@ import {
   navFactory,
   DetailsItem,
 } from '@console/internal/components/utils';
-import { labelForNodeKind, labelKeyForNodeKind } from '@console/shared';
 import { PodDisruptionBudgetModel } from '../../models';
 import { PodDisruptionBudgetKind } from './types';
 
@@ -20,51 +18,38 @@ const menuActions = [...Kebab.getExtensionsActionsForKind(PodDisruptionBudgetMod
 const PodDisruptionBudgetDetails: React.FC<PodDisruptionBudgetDetailsProps> = ({ obj }) => {
   const { t } = useTranslation();
   return (
-    <>
-      <Helmet>
-        <title data-title-id={`${labelForNodeKind(obj.kind)} · Details`}>
-          {obj.metadata.name}
-          {' · '} {t(labelKeyForNodeKind(obj.kind))}
-          {' · '} {t('console-app~Details')}
-        </title>
-      </Helmet>
-      <div className="co-m-pane__body">
-        <SectionHeading text={t('console-app~PodDisruptionBudget details')} />
-        <div className="co-m-pane__body-group">
-          <div className="row">
-            <div className="col-sm-6">
-              <ResourceSummary resource={obj} showPodSelector />
-            </div>
-            <div className="col-sm-6">
-              <dl className="co-m-pane__details">
-                <DetailsItem
-                  label={
-                    !_.isNil(obj.spec.minAvailable)
-                      ? t('console-app~Min available')
-                      : t('console-app~Max unavailable')
-                  }
-                  obj={obj}
-                  path={
-                    !_.isNil(obj.spec.minAvailable) ? 'spec.minAvailable' : 'spec.maxUnavailable'
-                  }
-                >
-                  {!_.isNil(obj.spec.minAvailable)
-                    ? obj.spec.minAvailable
-                    : obj.spec.maxUnavailable}
-                </DetailsItem>
-                <DetailsItem
-                  label={t('console-app~Allowed disruption')}
-                  obj={obj}
-                  path="status.disruptionsAllowed"
-                >
-                  {obj.status.disruptionsAllowed}
-                </DetailsItem>
-              </dl>
-            </div>
+    <div className="co-m-pane__body">
+      <SectionHeading text={t('console-app~PodDisruptionBudget details')} />
+      <div className="co-m-pane__body-group">
+        <div className="row">
+          <div className="col-sm-6">
+            <ResourceSummary resource={obj} showPodSelector />
+          </div>
+          <div className="col-sm-6">
+            <dl className="co-m-pane__details">
+              <DetailsItem
+                label={
+                  !_.isNil(obj.spec.minAvailable)
+                    ? t('console-app~Min available')
+                    : t('console-app~Max unavailable')
+                }
+                obj={obj}
+                path={!_.isNil(obj.spec.minAvailable) ? 'spec.minAvailable' : 'spec.maxUnavailable'}
+              >
+                {!_.isNil(obj.spec.minAvailable) ? obj.spec.minAvailable : obj.spec.maxUnavailable}
+              </DetailsItem>
+              <DetailsItem
+                label={t('console-app~Allowed disruption')}
+                obj={obj}
+                path="status.disruptionsAllowed"
+              >
+                {obj.status.disruptionsAllowed}
+              </DetailsItem>
+            </dl>
           </div>
         </div>
       </div>
-    </>
+    </div>
   );
 };
