+++ frontend/packages/console-app/src/components/volume-snapshot/volume-snapshot-content-details.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import Helmet from 'react-helmet';
 import { useTranslation } from 'react-i18next';
 import { ResourceEventStream } from '@console/internal/components/events';
 import { DetailsPage, DetailsPageProps } from '@console/internal/components/factory';
@@ -13,7 +12,7 @@ import {
 } from '@console/internal/components/utils';
 import { VolumeSnapshotClassModel, VolumeSnapshotModel } from '@console/internal/models';
 import { referenceForModel, VolumeSnapshotContentKind } from '@console/internal/module/k8s';
-import { labelForNodeKind, labelKeyForNodeKind, Status } from '@console/shared';
+import { Status } from '@console/shared';
 import { volumeSnapshotStatus } from '../../status';
 
 const { editYaml, events } = navFactory;
@@ -27,69 +26,60 @@ const Details: React.FC<DetailsProps> = ({ obj }) => {
   const sizeMetrics = size ? humanizeBinaryBytes(size).string : '-';
 
   return (
-    <>
-      <Helmet>
-        <title data-title-id={`${labelForNodeKind(obj.kind)} · Details`}>
-          {obj.metadata.name}
-          {' · '} {t(labelKeyForNodeKind(obj.kind))}
-          {' · '} {t('console-app~Details')}
-        </title>
-      </Helmet>
-      <div className="co-m-pane__body">
-        <SectionHeading text={t('console-app~VolumeSnapshotContent details')} />
-        <div className="row">
-          <div className="col-md-6 col-xs-12">
-            <ResourceSummary resource={obj}>
-              <dt>{t('console-app~Status')}</dt>
-              <dd>
-                <Status status={volumeSnapshotStatus(obj)} />
-              </dd>
-            </ResourceSummary>
-          </div>
-          <div className="col-md-6">
-            <dl className="co-m-pane__details">
-              {size && (
-                <>
-                  <dt>{t('console-app~Size')}</dt>
-                  <dd>{sizeMetrics}</dd>
-                </>
-              )}
-              <dt>{t('console-app~VolumeSnapshot')}</dt>
-              <dd>
-                <ResourceLink
-                  kind={referenceForModel(VolumeSnapshotModel)}
-                  name={snapshotName}
-                  namespace={snapshotNamespace}
-                />
-              </dd>
-              <dt>{t('console-app~VolumeSnapshotClass')}</dt>
-              <dd>
-                <ResourceLink
-                  kind={referenceForModel(VolumeSnapshotClassModel)}
-                  name={obj?.spec?.volumeSnapshotClassName}
-                />
-              </dd>
-              <dt>{t('console-app~Deletion policy')}</dt>
-              <dd>{deletionPolicy}</dd>
-              <dt>{t('console-app~Driver')}</dt>
-              <dd>{driver}</dd>
-              {volumeHandle && (
-                <>
-                  <dt>{t('console-app~Volume handle')}</dt>
-                  <dd>{volumeHandle}</dd>
-                </>
-              )}
-              {snapshotHandle && (
-                <>
-                  <dt>{t('console-app~Snapshot handle')}</dt>
-                  <dd>{snapshotHandle}</dd>
-                </>
-              )}
-            </dl>
-          </div>
+    <div className="co-m-pane__body">
+      <SectionHeading text={t('console-app~VolumeSnapshotContent details')} />
+      <div className="row">
+        <div className="col-md-6 col-xs-12">
+          <ResourceSummary resource={obj}>
+            <dt>{t('console-app~Status')}</dt>
+            <dd>
+              <Status status={volumeSnapshotStatus(obj)} />
+            </dd>
+          </ResourceSummary>
+        </div>
+        <div className="col-md-6">
+          <dl className="co-m-pane__details">
+            {size && (
+              <>
+                <dt>{t('console-app~Size')}</dt>
+                <dd>{sizeMetrics}</dd>
+              </>
+            )}
+            <dt>{t('console-app~VolumeSnapshot')}</dt>
+            <dd>
+              <ResourceLink
+                kind={referenceForModel(VolumeSnapshotModel)}
+                name={snapshotName}
+                namespace={snapshotNamespace}
+              />
+            </dd>
+            <dt>{t('console-app~VolumeSnapshotClass')}</dt>
+            <dd>
+              <ResourceLink
+                kind={referenceForModel(VolumeSnapshotClassModel)}
+                name={obj?.spec?.volumeSnapshotClassName}
+              />
+            </dd>
+            <dt>{t('console-app~Deletion policy')}</dt>
+            <dd>{deletionPolicy}</dd>
+            <dt>{t('console-app~Driver')}</dt>
+            <dd>{driver}</dd>
+            {volumeHandle && (
+              <>
+                <dt>{t('console-app~Volume handle')}</dt>
+                <dd>{volumeHandle}</dd>
+              </>
+            )}
+            {snapshotHandle && (
+              <>
+                <dt>{t('console-app~Snapshot handle')}</dt>
+                <dd>{snapshotHandle}</dd>
+              </>
+            )}
+          </dl>
         </div>
       </div>
-    </>
+    </div>
   );
 };
