+++ frontend/packages/console-shared/src/utils/utils.ts
@@ -1,8 +1,7 @@
 import { JSONSchema7 } from 'json-schema';
-import * as _ from 'lodash';
 import { toPath } from 'lodash';
 import { FirehoseResult } from '@console/internal/components/utils/types';
-import { K8sKind, K8sResourceKind, modelFor } from '@console/internal/module/k8s';
+import { K8sResourceKind } from '@console/internal/module/k8s';
 import { getUID } from '../selectors/common';
 
 export type EntityMap<A> = { [propertyName: string]: A };
@@ -89,31 +88,4 @@ export const alphanumericCompare = (a: string, b: string): number => {
     numeric: true,
     sensitivity: 'base',
   });
-};
-
-export const translationForResourceKind = {
-  // t('console-shared~Helm Release')
-  HelmRelease: `console-shared~Helm Release`,
-};
-
-export const labelForNodeKind = (kindString: string) => {
-  const model: K8sKind | undefined = modelFor(kindString);
-  if (model) {
-    return model.label;
-  }
-  return _.startCase(kindString);
-};
-
-export const labelKeyForNodeKind = (kindString: string) => {
-  const model: K8sKind | undefined = modelFor(kindString);
-  if (model) {
-    if (model.labelKey) {
-      return model.labelKey;
-    }
-    return model.label;
-  }
-  if (translationForResourceKind[kindString]) {
-    return translationForResourceKind[kindString];
-  }
-  return _.startCase(kindString);
 };
