+++ frontend/packages/console-telemetry-plugin/src/listeners/segment.ts
@@ -106,12 +106,7 @@ export const eventListener: TelemetryEventListener = async (
       }
       break;
     case 'page':
-      {
-        const titleElement = document.querySelector('title');
-        const titleValue = titleElement.getAttribute('data-title-id');
-        properties.title = titleValue || document.title;
-        (window as any).analytics.page(undefined, properties, anonymousIP);
-      }
+      (window as any).analytics.page(undefined, properties, anonymousIP);
       break;
     default:
       (window as any).analytics.track(eventType, properties, anonymousIP);
