+++ frontend/packages/container-security/locales/en/container-security.json
@@ -1,6 +1,5 @@
 {
   "Vulnerabilities": "Vulnerabilities",
-  "Details": "Details",
   "Image Manifest Vulnerabilities details": "Image Manifest Vulnerabilities details",
   "Registry": "Registry",
   "Manifest": "Manifest",
