+++ frontend/packages/operator-lifecycle-manager/locales/en/olm.json
@@ -223,7 +223,6 @@
   "Resource Kind": "Resource Kind",
   "All Instances": "All Instances",
   "Create new": "Create new",
-  "Details": "Details",
   "{{kind}} overview": "{{kind}} overview",
   "{{item}} details": "{{item}} details",
   "Resources": "Resources",
