+++ frontend/packages/topology/src/components/list-view/ListElementWrapper.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { Node, isNode } from '@patternfly/react-topology';
 import { observer } from 'mobx-react';
-import { labelForNodeKind } from '@console/shared';
 import { getResourceKind } from '../../utils/topology-utils';
+import { labelForNodeKind } from './list-view-utils';
 import { listViewNodeComponentFactory } from './listViewComponentFactory';
 
 interface ListElementWrapperProps {
