+++ frontend/packages/topology/src/components/list-view/TopologyListViewKindGroup.tsx
@@ -9,7 +9,7 @@ import {
 } from '@patternfly/react-core';
 import { GraphElement, isNode, Node, observer } from '@patternfly/react-topology';
 import { useTranslation } from 'react-i18next';
-import { labelKeyForNodeKind } from '@console/shared';
+import { labelKeyForNodeKind } from './list-view-utils';
 import ListElementWrapper from './ListElementWrapper';
 
 interface TopologyListViewKindGroupProps {
