+++ frontend/packages/topology/src/components/list-view/list-view-utils.tsx
@@ -1,6 +1,34 @@
 import { isNode, Node, GraphElement } from '@patternfly/react-topology';
-import { labelForNodeKind } from '@console/shared';
+import * as _ from 'lodash';
+import { K8sKind, modelFor } from '@console/internal/module/k8s';
 import { getResourceKind } from '../../utils/topology-utils';
+
+export const translationForResourceKind = {
+  // t('topology~Helm Release')
+  HelmRelease: `topology~Helm Release`,
+};
+
+export const labelForNodeKind = (kindString: string) => {
+  const model: K8sKind | undefined = modelFor(kindString);
+  if (model) {
+    return model.label;
+  }
+  return _.startCase(kindString);
+};
+
+export const labelKeyForNodeKind = (kindString: string) => {
+  const model: K8sKind | undefined = modelFor(kindString);
+  if (model) {
+    if (model.labelKey) {
+      return model.labelKey;
+    }
+    return model.label;
+  }
+  if (translationForResourceKind[kindString]) {
+    return translationForResourceKind[kindString];
+  }
+  return _.startCase(kindString);
+};
 
 export const getChildKinds = (children: GraphElement[]) => {
   const childNodes = children.filter((n) => isNode(n)) as Node[];
