+++ frontend/public/locales/en/public.json
@@ -425,6 +425,7 @@
   "Service mesh enabled": "Service mesh enabled",
   "Inventory": "Inventory",
   "Launcher": "Launcher",
+  "Project overview": "Project overview",
   "ResourceQuotas": "ResourceQuotas",
   "AppliedClusterResourceQuotas": "AppliedClusterResourceQuotas",
   "No AppliedClusterResourceQuotas": "No AppliedClusterResourceQuotas",
@@ -450,7 +451,6 @@
   "Min ready seconds": "Min ready seconds",
   "Triggers": "Triggers",
   "DeploymentConfig details": "DeploymentConfig details",
-  "Environment": "Environment",
   "ReplicationControllers": "ReplicationControllers",
   "Retry rollout": "Retry rollout",
   "DeploymentConfigs": "DeploymentConfigs",
@@ -1240,6 +1240,7 @@
   "Memory usage by pod (top 10)": "Memory usage by pod (top 10)",
   "Resource usage": "Resource usage",
   "NetworkPolicies": "NetworkPolicies",
+  "Project details": "Project details",
   "RoleBindings": "RoleBindings",
   "Target pods": "Target pods",
   "To ports": "To ports",
@@ -1330,7 +1331,6 @@
   "The terminal connection has closed.": "The terminal connection has closed.",
   "Reconnect": "Reconnect",
   "Connecting to": "Connecting to",
-  "Logs": "Logs",
   "Receiving Traffic": "Receiving Traffic",
   "Not Receiving Traffic": "Not Receiving Traffic",
   "Log in to your account": "Log in to your account",
@@ -1356,7 +1356,6 @@
   "Image pull secret": "Image pull secret",
   "Pod details": "Pod details",
   "Init containers": "Init containers",
-  "Terminal": "Terminal",
   "Running": "Running",
   "CrashLoopBackOff": "CrashLoopBackOff",
   "Failed": "Failed",
@@ -1723,7 +1722,10 @@
   "Non-printable file detected.": "Non-printable file detected.",
   "File contains non-printable characters. Preview is not available.": "File contains non-printable characters. Preview is not available.",
   "Deleting": "Deleting",
+  "Logs": "Logs",
+  "Environment": "Environment",
   "Cluster Operators": "Cluster Operators",
+  "Terminal": "Terminal",
   "Delete {{kind}}": "Delete {{kind}}",
   "Edit {{kind}}": "Edit {{kind}}",
   "Edit Pod selector": "Edit Pod selector",
